Remove table component
