+++ src/components/DepositWidget/index.tsx
@@ -1,7 +1,10 @@
-import Widget from '../layout/Widget'
 import React from 'react'
 import styled from 'styled-components'
-import Table from './Table'
+
+import Widget from '../layout/Widget'
+import { Row } from './Row'
+import { tokenApi } from 'api'
+import { Network, TokenBalanceDetails } from 'types'
 
 const Wrapper = styled.div`
   display: flex;
@@ -11,18 +14,51 @@ const Wrapper = styled.div`
     align-self: flex-end;
   }
 `
+
+function _getBalances(): TokenBalanceDetails[] {
+  const tokens = tokenApi.getTokens(Network.Rinkeby)
+  // Mock implementation
+  return tokens.map(token => ({
+    ...token,
+    exchangeWallet: 0,
+    pendingDeposits: 0,
+    pendingWithdraws: 0,
+    enabled: true,
+  }))
+}
+
 // TODO: add correct source address
 const contractSource = 'https://gnosis.io'
 
-const DepositWidget: React.FC = () => (
-  <Widget>
-    <Wrapper>
-      <a href={contractSource} className="contractLink">
-        View verified contract
-      </a>
-      <Table />
-    </Wrapper>
-  </Widget>
-)
+const DepositWidget: React.FC = () => {
+  const tokenBalancesList = _getBalances()
+  return (
+    <Widget>
+      <Wrapper>
+        <a href={contractSource} className="contractLink">
+          View verified contract
+        </a>
+        <table>
+          <thead>
+            <tr>
+              <th></th>
+              <th>Token</th>
+              <th>Exchange wallet</th>
+              <th>Pending deposits</th>
+              <th>Pending withdraws</th>
+              <th></th>
+              <th></th>
+            </tr>
+          </thead>
+          <tbody>
+            {tokenBalancesList.map(tokenBalances => (
+              <Row key={tokenBalances.addressMainnet} tokenBalances={tokenBalances} />
+            ))}
+          </tbody>
+        </table>
+      </Wrapper>
+    </Widget>
+  )
+}
 
 export default DepositWidget
